+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Product Backlog
-* [Bug] Teachers should not be able to get email, phone, or google in user.controller.js:index()
 
 <a name="1.0.17"></a>
+* [Bug] Teachers should not be able to get email, phone, or google in user.controller.js:index()
 * Improve login page - validation, separate Google and SHY logins
 * Add toast after forgotpassword and message (angularjs-toaster)
 * [Bug] usereditor.pug - Teachers should not see phone field.